rename backend to stats-api
