+++ stats-project/tests/docker-compose/docker-compose.yaml
@@ -15,7 +15,7 @@ services:
     restart: unless-stopped
 
   mysql:
-    image: mysql:5.7
+    image: mysql:8.0
     container_name: test-mysql
     environment:
       MYSQL_ROOT_PASSWORD: root_password
